+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - CI: Verified support on Python 3.13
+- Dependencies: Updated to `crate-1.0.0`
 
 ## 2024/10/07 0.40.0
 - Propagate error traces properly, using the `error_trace` `connect_args` option,